requirements.txt
blinker==1.7.0
boto3==1.29.6
botocore==1.32.6
certifi==2023.7.22
chardet==3.0.4
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.0
idna==2.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
jdcal==1.4.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.5
MarkupSafe==2.1.3
numpy==1.19.2
openpyxl==3.0.5
pandas==1.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.24.0
s3transfer==0.7.0
six==1.16.0
urllib3==1.25.11
Werkzeug==3.0.1
xlrd==1.2.0
XlsxWriter==3.1.9
zipp==3.17.0
